pyproject.toml
[tool.poetry]
name = "api automation"
version = "0.1.0"
description = "Upwork API coding challenge"
authors = ["Dr. Nishanthan Kamaleson"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
pygithub = "^2.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
